backend/requirements.txt
Flask
python-dotenv
PyPDF2
Pillow
pytesseract
sentence-transformers
faiss-cpu
requests
